@@ -1,8 +1,9 @@
-"use strict";
-exports.__esModule = true;
-exports.getLogin = function (req, res) {
-    res.render('login.ejs');
+/*
+export const getLogin = function (req, res) {
+  res.render('login.ejs');
 };
-exports.postLogin = function (req, res) {
-    res.render('login.ejs');
+
+export const postLogin = function (req, res) {
+  res.render('login.ejs');
 };
+*/ 
