@@ -11,20 +11,8 @@ export default class Organization {
 
   private orgAdminIds: number[];
 
-  private orgWorkerIds: number[];
-
-  private orgClientIds: number[];
-
-  private orgWorkers: OrgWorker[];
-
-  private orgClients: Client[];
-
   // getter and setter methods
   constructor() {
-    this.orgAdminIds = [];
-    this.orgWorkerIds = [];
-    this.orgClients = [];
-    this.orgWorkers = [];
     this.createId();
   }
 
@@ -37,24 +25,6 @@ export default class Organization {
 
   getAdminIds() : number[] { return this.orgAdminIds; }
 
-  getWorkerIds() : number[] { return this.orgWorkerIds; }
-
-  getWorkers() : OrgWorker[] { return this.orgWorkers; }
-
-  getClients() : Client[] { return this.orgClients; }
-
-  getClientIds() : number[] { return this.orgClientIds; }
-
-
-  addWorker(orgWorker : OrgWorker) : void {
-    this.orgWorkers.push(orgWorker);
-    this.orgWorkerIds.push(orgWorker.getId());
-  }
-
-  addClient(client : Client) : void {
-    this.orgClients.push(client);
-    this.orgClientIds.push(client.getId());
-  }
 
   createId() : void {
     const json = JSON.parse(readFileSync('./counter.json', 'utf8'));
@@ -68,9 +38,7 @@ export default class Organization {
       orgId: this.orgId,
       orgName: this.orgName,
       orgDescription: this.orgDescription,
-      orgAdminIds: this.orgAdminIds,
-      orgWorkerIds: this.orgWorkerIds,
-      orgClientsIds: this.orgClientIds,
+      orgAdminIds: this.orgAdminIds
     });
   }
 }
